Add files via upload
In this first project, the basic fundamentals taught to us thus far are expressed in this interactive experience. The user can move, click, and press their way to a new abstract experience one loop at a time. The basic syntax are exploited through randomness and the randomness can only happen with the user interacts with the domain.